CA-77 added the AuthProvider to wrap the app
