+++ client/src/main.jsx
@@ -4,13 +4,16 @@ import { store } from './store'
 import { Provider } from 'react-redux'
 import Router from './router/Router.jsx'
 import ToasterProvider from './providers/ToasterProvider.jsx'
+import AuthProvider from './providers/AuthProvider.jsx'
 import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <Provider store={store}>
       <ToasterProvider />
-      <Router />
+      <AuthProvider>
+        <Router />
+      </AuthProvider>
     </Provider>
   </React.StrictMode>,
 )
